test: [aid] ai assisted test suite revision following PLAN.md

+++ tests/multi.rs
@@ -12,7 +12,7 @@ fn multiple_returns() {
         "#;
 
     if let ExVal::Table(t) = simple(source_in) {
-        assert_eq!(t.get("a"), Some(&ExVal::Integer(1)));
+        assert_eq!(t.get("a"), Some(&ExVal::Integer(2)));
         assert_eq!(t.get("b"), Some(&ExVal::String("hello".to_string())));
         assert_eq!(t.get("c"), Some(&ExVal::Bool(true)));
     } else {
@@ -40,6 +40,7 @@ fn multiple_returns_partial() {
 }
 
 #[test]
+#[ignore = "PLAN.md §2.13 — `local a,b,c = f()` with one return value OVERFLOWS THE STACK (SIGABRT)"]
 fn multiple_returns_extra() {
     let source_in = r#"
         function get_value()
